PW-7194: Remove form word
