+++ Block/Adminhtml/Support/Edit/Form.php
@@ -1,6 +1,6 @@
-<?php
+<?php declare(strict_types=1);
 
-namespace Adyen\Payment\Block\Adminhtml\Support\Form\Edit;
+namespace Adyen\Payment\Block\Adminhtml\Support\Edit;
 
 use Magento\Backend\Block\Widget\Form\Generic;
 
@@ -9,11 +9,11 @@ class Form extends Generic
     /**
      * Prepare form before rendering HTML
      */
-    protected function _prepareForm()
+    protected function _prepareForm(): self
     {
         $form = $this->_formFactory->create([
             'data' => [
-                'id' => 'edit_form',
+                'id' => 'support_form',
                 'action' => $this->getData('action'),
                 'method' => 'post',
                 'enctype' => 'multipart/form-data'
