+++ Controller/Adminhtml/Configuration/Support.php
@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace Adyen\Payment\Controller\Adminhtml\Configuration;
 
@@ -11,7 +11,7 @@ class Support extends Action
     {
         $resultPage = $this->resultFactory->create(ResultFactory::TYPE_PAGE);
         $resultPage->setActiveMenu('Adyen_Payment::configuration_support')
-            ->getConfig()->getTitle()->prepend(__('Adyen Support Form'));
+            ->getConfig()->getTitle()->prepend(__('Adyen Support'));
         return $resultPage;
     }
 }
