+++ view/adminhtml/layout/adyen_configuration_support.xml
@@ -1,16 +1,16 @@
 <?xml version="1.0"?>
 <page layout="admin-2columns-left" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
     <head>
-        <title>Adyen Support Form</title>
+        <title>Adyen Support</title>
     </head>
     <body>
         <referenceContainer name="content">
-            <block class="Adyen\Payment\Block\Adminhtml\Edit" name="support_form"/>
+            <block class="Adyen\Payment\Block\Adminhtml\Support\Edit" name="support_form"/>
         </referenceContainer>
         <referenceContainer name="left">
-            <block class="Adyen\Payment\Block\Adminhtml\Support\Form\Edit\Tabs" name="support_form_tabs">
-                <block class="Adyen\Payment\Block\Adminhtml\Support\Form\Edit\Tab\OrderProcessing" name="support_form_tab_order"/>
-                <block class="Adyen\Payment\Block\Adminhtml\Support\Form\Edit\Tab\ConfigurationSettings" name="support_form_tab_config"/>
+            <block class="Adyen\Payment\Block\Adminhtml\Support\Edit\Tabs" name="support_form_tabs">
+                <block class="Adyen\Payment\Block\Adminhtml\Support\Edit\Tab\OrderProcessing" name="support_form_tab_order"/>
+                <block class="Adyen\Payment\Block\Adminhtml\Support\Edit\Tab\ConfigurationSettings" name="support_form_tab_config"/>
                 <action method="addTab">
                     <argument name="name" xsi:type="string">order_section</argument>
                     <argument name="block" xsi:type="string">support_form_tab_order</argument>
