+++ view/adminhtml/templates/form/support.phtml
@@ -5,7 +5,7 @@
 require(['jquery'], function($) {
     $('#send').click(() => {
         // force submit without validation
-        $('#edit_form')[0].submit();
+        $('#support_form')[0].submit();
     });
 });
 </script>
